sales report controller added, separated sales report api's from invoice api
