+++ routes/index.js
@@ -7,6 +7,7 @@ const userRoutes = require('./userRoutes');
 const stockRoutes = require('./stockRoutes');
 const invoiceRoutes = require('./invoiceRoutes');
 const customerRoutes = require('./customerRoutes');
+const salesReportRoutes = require('./salesReport');
 
 // Combine all routes with their specific prefixes
 router.use('/v1/auth', authRoutes);
@@ -14,5 +15,6 @@ router.use('/v1/user', userRoutes);
 router.use('/v1/stock', stockRoutes);
 router.use('/v1/invoice', invoiceRoutes);
 router.use('/v1/customer', customerRoutes);
+router.use('/v1/sales-report', salesReportRoutes);
 
 module.exports = router;
