+++ routes/invoiceRoutes.js
@@ -8,12 +8,6 @@ const {
 	deleteInvoice,
 	generatePDF,
 	searchInvoices,
-	getSalesByPrice,
-	getSalesByQuantity,
-	getMonthlySales,
-	getYearlySales,
-	getDailySalesForMonth,
-	getProductSalesForMonth
 } = require('../controllers/invoiceController');
 const {
 	authenticateUser,
@@ -25,15 +19,6 @@ router.post('/', authenticateUser, createInvoice);
 
 // Get all invoices with pagination and filtering
 router.get('/', authenticateUser, getAllInvoices);
-
-// Sales Report Routes
-router.get('/reports/by-price', authenticateUser, getSalesByPrice);
-router.get('/reports/by-quantity', authenticateUser, getSalesByQuantity);
-// router.get('/reports/monthly', authenticateUser, getMonthlySales);
-router.get('/reports/yearly', authenticateUser, getYearlySales);
-router.get('/reports/daily', authenticateUser, getDailySalesForMonth);
-router.get('/reports/product-sales', authenticateUser, getProductSalesForMonth);
-
 
 // Search invoices by date range and customer
 // Query parameters:
